requirements.txt
bcrypt
blinker
certifi
cffi
click
Flask
Flask-Bcrypt
Flask-Login
Flask-Mail
Flask-SQLAlchemy
Flask-WTF
itsdangerous
Jinja2
MarkupSafe
Pillow
pycparser
six
SQLAlchemy
Werkzeug
WTForms
email-validator
flask-share
gunicorn
pygeoip
requests
pdfkit
stripe
pyopenssl
cryptography
configobj
wkhtmltopdf